UI Changes
I updated the layout of the app when opened. I changed Header items to
be in black font, to offset the see of white. This makes it more
pleasing to the eye and easier to scroll through when the fields are
blank.

I also lowered the Alpha of the dictation and submit buttons so they
don't full obstruct from the view of the app fields.

I also changed the colors of the background, dictate, and submit
buttons. These are more neutral in color, so the app feels more natural,
and real.

I also change the app launcher icon and also added a copy to the
background of the app.

If anyone has any objections to these changes I will change them.