Lots of changes/additions/updates read Description
Warning: It may look like I deleted your code, but I didn't. I had to
remove and add your code in if that's the case. In reality no code was
changed not done by myself.

For starters, I changed the User Interface more, by using more
refreshing buttons. In addition I also added a submit button to the top
menu bar, and also fixed the issue where the background would resize
when the keyboard raised and lowered. This caused an issue where now the
bottom buttons will be hidden when the keyboard raises. This bug was
added to the known bugs list (more below).

In addition to UI Changes I also added a HIPPA field to the data base so
we can store the value of HIPPA Waiver in the database. This way we can
fix the landscape mode bugs with the waiver. To then ultimately
re-implement landscape support.

A Continuous Dictation file was added to the data base, but the code
isn't ready to be added from the model repository. We hope to get this
move over by Monday if all goes well.

A new thing I added was a KnownBugs.txt to the project. Please use this
to announce any bugs you may have found so someone can attempt to fix
them. Just be sure to update the list if you fix one as well.

Lastly I performed a little house cleaning and removed some retired
pictures, and code lines I found.  As well as updating the gradle file
to 1.1, from 1.0. So if your project won't build because of a gradle
error, you need to update your Android Studio IDE.

Sorry for all the change, but I had some bugs that prevented me from
pushing earlier.