Background fix
Forgot to push the manifest file update that fixes the background
becoming distorted when you raise and lower the keyboard. Also I
changed the name of the app from Medical Fax to just simply Medical
App. Considering it won’t actually fax, I don’t see a need to label it
as such.