Beginning of Continuous Dictation Implementation
Code for continuous dictation was added and comments were included.

Currently, the result of dictation is not passed to the presenter class
(since it is auto added to the database). Since it is auto added to the
database, presenter will need to do a check for continuous dictation in
startTranscription(). If it is enabled, they can then load the fields
from the database after calling getSpeechInput().

I added some comments in the presenter class on how this could be done.

Right now continuous dictation is not functional.