Icon active states
I added all the buttons to the action bar, to hot fix the keyboard bug.
So for now the user will have to use the action bar icons when the
keyboard is raised. Working on the bottom screen action bar to remedy
that issue.

I also added two states to the icons, so when the user presses a button,
they get a visual response.

I also started working on a 3 icon format for the dictation button. 1
icon to display a no microphone available state, a normal state
(currently in use), and an active state (currently in use).

To get this to work, add a variable for the availability of dictation in
the database, as well allowed it to be updated if a microphone becomes
available. This feature is not fully finished yet.